* fix: close pull request (#751)

* Fix: publish button delay (#752)

* Fix: publish button delay

* Fix: tests

* Fix: move review request status check into router

* feat: return sites from db for email login (#754)

* feat: return sites from db for email login

* fix: tests

* fix: use lodash function for type guard

* fix(githhub service): get call to github to prevent race condition (#756)

* fix(githhub service): get call to github to prevent race condition

* fix(github service): throw error if directory name not defined

* style(github service): add comments for clarity

* fix: tests

---------


* Fix: handle updating of files in root directory (#761)

* fix: create .keep first when rename subfolder (#762)

* 0.26.0

---------

+++ src/routes/v2/authenticated/review.ts
@@ -716,13 +716,19 @@ export class ReviewsRouter {
       return res.status(404).json({ message: possibleReviewRequest.message })
     }
 
-    // Step 4: Merge review request
+    // Step 4: Check if review request is valid
+    if (possibleReviewRequest.reviewStatus !== ReviewRequestStatus.Approved)
+      return res
+        .status(404)
+        .json({ message: "Approved review request not found!" })
+
+    // Step 5: Merge review request
     // NOTE: We are not checking for existence of PR
     // as the underlying Github API returns 404 if
     // the requested review could not be found.
     await this.reviewRequestService.mergeReviewRequest(possibleReviewRequest)
 
-    // Step 5: Clean up the review request view records
+    // Step 6: Clean up the review request view records
     // The error is discarded as we are guaranteed to have a review request
     await this.reviewRequestService.deleteAllReviewRequestViews(
       site.value,
